chore: Remove boilerplate configs
This removes all of the boilerplate configurations that are not needed
for a module. I'm not sure what these are used for but its quite a bit
of overhead for a Terraform module. Instead, it has been replaced with a
more simplified setup based on what is used in terraform-aws-modules and
EKS Blueprints modules. This includes additional pre-commit checks used
for development workflow as well as an associated set of CI checks to
ensure that future contributions are aligning to the pre-commit check
requirements.

The terratest setup was largely un-used so it has been removed but can
always be re-introduced in the future if a more robust set of tests will
be configured with that setup. For now, lets keep it to the bare minimum
of what is required and being used in the project. Less is more, more is
less!