fix: Use `this` naming convention for resource names
Its convention within modules to use `this` for the resource name when
there is only one of those resources within the module scope. The name
is not too important and just sort of shows the intent that it just
means "this module". When you have multiple resources in the module
scope, then more distinct and descriptive names should be used - but
keep these names short since they will be shown in output to users
(plans and applies).