fix: Ensure variable names are consistent with the API
The variable names used should be very consitent with the API so that
users do not need to re-learn the module's way of specifying values
(unless there is a strong reason to deviate). There are two reasons
where deviating can help:
1. When using loops, its helpful to make the variable plural to give
   users the hint that multiple values can be provided
2. When the API has some inconsitencies (rare, but does happen). There
   is a great example of you doing this here where the API has two very
similar names but they are not consistent
`enable_auto_branch_creation` and `enable_branch_auto_deletion`. Making
these more consistent and aligned with one another should help users
(its not always perfect because there are those that look only at the
API and then will ask why the deviation)