chore: Reduce the amount of redundant examples
It is easy to create many different examples due to the various
configuration possibilities. However, over time this is difficult to
maintain. Instead, try to select a few examples taht broadly cover the
various configurations. At minimum, the entirety of the examples should
strive to cover all of the possible variables and arguments so that you
have a way to both test and demonstrate how to enable a certain
configuration.