chore: Align `main.tf`, `variables.tf` and `outputs.tf`
For most modules, its customary to define all of the resources and data
sources in the `main.tf` and you can create segmentation with a bit of
blocked doc comments to separate the file. This makes it easy to scan
across the implementation to the variables and outputs to make sure
everything is aligned and also for future contributors to know what
variables/outputs are used in which locations. If the module is rather
large, you can look at splitting up `main.tf`, but try to avoid lots of
small files as this makes it tedious to traverse the codebase as you
have to move across multiple files to understand the relationships and
interactions.

Also, data sources are used to support the resource configurations.
These should always be located close to where they are used. If they are
commonly used across the codebase, you can place these at the top of
`main.tf`. The intent is to be able to quickly and easily look at the
code and understand whats involved with creating that resource.