fix: Correct and add output values
Enure all output values are provided for each resource; these are
typically those values found under the `Attribute Reference` of the
resources documentation page. Note - not all of these values need to be
re-exported as outputs, only those that are useful to downstream
resources/modules as input

When writing outputs, the output names should be in reference to the
name of the module. This module is named `amplify-app` which means that
the `aws_amplify_app` resource outputs should not re-state the name
since its already in context from the module name. So instead of an
output named `amplify_app_arn` which the long form would look like
`module.amplify_app.amplify_app_arn` - you should instead use just the
attribute name `arn` which when used in the long form is simply
`module.amplify_app.arn`. For the supporting resources, they should
include at least a portion of their resource name since those are not
within the context of the module name. For example - the output name for
the ARN of `aws_amplify_domain_association` can be
`domain_association_arn` which is then
`module.amplify_app.domain_assocaition_arn`. Working backwards from the
long form, it should read somewhat like a descriptive sentence without
repeating the same information.